# openapi.yaml
openapi: 3.0.0
info:
  title: task-api
  description: >-
    Task management REST API. Users own projects; projects contain tasks; access
    is scoped to the authenticated user at every layer.
  version: 0.1.0
components:
  schemas:
    def-0:
      description: A project
      type: object
      required:
        - id
        - userId
        - name
        - description
        - createdAt
        - updatedAt
      properties:
        id:
          type: integer
        userId:
          type: integer
        name:
          type: string
        description:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      title: projectResponseSchema
...
paths:
...
  /projects:
    get:
      responses:
        '200':
          description: A list of projects
          content:
            application/json:
              schema:
                description: A list of projects
                type: array
                items:
                  $ref: '#/components/schemas/def-0'
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              additionalProperties: false
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                description:
                  type: string
                  nullable: true
      responses:
        '201':
          description: A project
          content:
            application/json:
              schema:
                description: A project
                type: object
                required:
                  - id
                  - userId
                  - name
                  - description
                  - createdAt
                  - updatedAt
                properties:
                  id:
                    type: integer
                  userId:
                    type: integer
                  name:
                    type: string
                  description:
                    type: string
                    nullable: true
                  createdAt:
                    type: string
                  updatedAt:
                    type: string
  /projects/{id}:
    get:
      parameters:
        - schema:
            type: string
            pattern: ^[0-9]{1,15}$
          in: path
          name: id
          required: true
      responses:
        '200':
          description: A project
          content:
            application/json:
              schema:
                description: A project
                type: object
                required:
                  - id
                  - userId
                  - name
                  - description
                  - createdAt
                  - updatedAt
                properties:
                  id:
                    type: integer
                  userId:
                    type: integer
                  name:
                    type: string
                  description:
                    type: string
                    nullable: true
                  createdAt:
                    type: string
                  updatedAt:
                    type: string
    patch:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 100
                description:
                  type: string
                  nullable: true
      parameters:
        - schema:
            type: string
            pattern: ^[0-9]{1,15}$
          in: path
          name: id
          required: true
      responses:
        '200':
          description: A project
          content:
            application/json:
              schema:
                description: A project
                type: object
                required:
                  - id
                  - userId
                  - name
                  - description
                  - createdAt
                  - updatedAt
                properties:
                  id:
                    type: integer
                  userId:
                    type: integer
                  name:
                    type: string
                  description:
                    type: string
                    nullable: true
                  createdAt:
                    type: string
                  updatedAt:
                    type: string
    delete:
      parameters:
        - schema:
            type: string
            pattern: ^[0-9]{1,15}$
          in: path
          name: id
          required: true
      responses:
        '204':
          description: No content
...